fastapi/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
psycopg2-binary
pusher
requests
